second commit

+++ src/main/kotlin/structures/ExtendedTrade.kt
@@ -1,6 +1,7 @@
 package structures
 
 import formats.Tag
+import java.math.BigDecimal
 import java.time.LocalDateTime
 
 data class ExtendedTrade(
@@ -8,8 +9,8 @@ data class ExtendedTrade(
     override val direction : Direction,
     override val dateTime: LocalDateTime,
     override val itemID: String,
-    override val price: Double,
-    override val quantity: Int,
+    override val price: BigDecimal,
+    override val quantity: Long,
     override val buyer: String,
     override val seller: String,
     val nestedParameters: List<Tag>
