+++ src/main/kotlin/structures/Footer.kt
@@ -1,6 +1,6 @@
 package structures
 
 data class Footer(
-    val tradeCount: Int,
-    val tradesCharsCount: Int? = null
+    val tradeCount: Long,
+    val tradesCharsCount: Long? = null
 )
